funds-manager: execution_client: venues: cowswap: add API type definitions (#326)
